crud

+++ src/main/java/com/board/bootboard/domain/repository/BoardRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.board.bootboard.domain.entity.BoardEntity;
 
-public interface BoardRepository extends JpaRepository<BoardEntity, Long> {
+public interface BoardRepository extends JpaRepository<BoardEntity, Integer> {
     
 }
